% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cleanCorpus}
\alias{cleanCorpus}
\title{cleanCorpus}
\usage{
cleanCorpus(data, lower = T, removeP = T, removeWS = F, removeSW = F,
  multipattern = NULL, stem = F, removeN = F)
}
\arguments{
\item{data}{data for cleansing, data frame or character vector}

\item{lower}{a logical value indicating whether the character vectors translate characters in particular from upper to lower case or vice versa. (default : T)}

\item{removeP}{a logical value indicating whether the character vectors remove all punctuations. (default : T)}

\item{removeWS}{a logical value indicating whether the character vectors remove all white spaces. (default : F)}

\item{multipattern}{a character vector indicating the character vectors remove pattern. (default : NULL)}

\item{stem}{a logical value indicating whether the character vectors stemming characters. (default : F)}

\item{removeN}{a logical value indicating whether the character vectors remove all numbers. (default : T)}
}
\description{
Convert data data frame to Corpus
}
\examples{
stopwords=c('from','and','for','was','had','ncp','will','its','required','unit','require')
cleanCorpus(VoC.text, stem=T, mystopwords=stopwords)
}

