% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{combineIndices}
\alias{combineIndices}
\title{Combine Point and Track OIDs}
\usage{
combineIndices(in.pt, in.tk, by, out.type = "list", verbose = TRUE)
}
\arguments{
\item{in.pt}{character. File paths to the point data}

\item{in.tk}{character. File paths to the track data. This should be the same length as in.pt.}

\item{by}{character. Character vector of column names used to summarize the track data.}

\item{out.type}{character. Should the output be a list of data.tables by file or a single combined file (dt)}

\item{verbose}{logical. Should the input tables be printed}
}
\value{
either a list or data.table containing the by columns and the OID.POINT and OID.TRACK columns
}
\description{
Combine object IDs from point data and track data based on the columns used for track-level summaries
}
\details{
This function does not confirm that point and track files are paired other than by the merge.
This could produce hidden errors if point-level and track-level data are not in the same order
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\concept{data organization}
\concept{hmd}
\keyword{manip}
