% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{createBldgTiles}
\alias{createBldgTiles}
\title{Download building tiles}
\usage{
createBldgTiles(
  url =
    "https://minedbuildings.z5.web.core.windows.net/global-buildings/dataset-links.csv",
  tiles,
  run = "all",
  out_gpkg = "MS_buildings_WGS84.gpkg"
)
}
\arguments{
\item{url}{character. URL where the csv file of the building tiles are stored.
This defaults to "https://minedbuildings.z5.web.core.windows.net/global-buildings/dataset-links.csv"}

\item{tiles}{sf object. An sf object of the polygons of the tiles used to split the building data.}

\item{run}{numeric or "all". Which tiles should be included in the tile creation? Defaults to 'all' which includes all tiles in "tiles". Useful if the function fails and needs to be restarted from where it was left off.}

\item{out_gpkg}{character. File path of the geopackage to save the output tiles.}
}
\value{
A list of file paths pointing to the output location of the tiles
}
\description{
Download building tiles from Microsoft's building detector
}
\details{
See https://github.com/microsoft/GlobalMLBuildingFootprints#will-there-be-more-data-coming-for-other-geographies for details on how the building detector works.
This function simply downloads already created data and separates it into the USGS tiles used for classification.
Currently, this function will redownload all the data and not just updates.
Microsoft's current system for defining updates do not state where updates come from so it is impossible to know what the most recent data is.
Running this function takes about 1 week for the Contiguous United States (CONUS)
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\concept{buildings}
\concept{data download}
\keyword{datasets}
\keyword{manip}
