% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{distToFeature}
\alias{distToFeature}
\title{Distance to the nearest feature}
\usage{
distToFeature(x, n, feature)
}
\arguments{
\item{x}{sf object. Spatial data.frame of the input data}

\item{n}{String. Name that should be used for output files}

\item{feature}{sf object. Spatial data.frame of the nearest feature}
}
\value{
a data.table object in the same order as X with the distance to nearest feature and features of the nearest feature.
}
\description{
Simplified calculation of distance to nearest feature while maintaining structure of nearest feature data
}
\details{
This is just a simple function to calculate distance to nearest feature and add features of the nearest feature to the output.
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\seealso{
\code{\link{classifyHMD}}
}
\concept{Distance to Feature}
\concept{HMD}
\keyword{methods}
