% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_functions.R
\name{flaggedToRaw}
\alias{flaggedToRaw}
\title{Flagged data to raw data}
\usage{
flaggedToRaw(in.dir, out.dir)
}
\arguments{
\item{in.dir}{character. Directory containing the flagged HMD data}

\item{out.dir}{character. The directory where the raw data should be saved}
}
\value{
Vector of file paths leading to the newly (re)created raw data
}
\description{
Convert flagged data back to raw (a reverse of flagAssignment)
}
\details{
This function takes flagged data and converts it back to raw (by removing columns created by the flagging process).
Beware, if using this function on cleaned data, data removed during the cleaning process will not be restored.
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\seealso{
\code{\link{flagAssignment}}, \code{\link{flagRemoval}}
}
\concept{HMD}
\keyword{manip}
