% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_functions.R
\name{flagRemoval}
\alias{flagRemoval}
\title{Remove data from flagged HMD}
\usage{
flagRemoval(
  in.dir,
  out.dir,
  FF.remove = NULL,
  FF.suspect = NULL,
  method = "none",
  all.keep = FALSE,
  move.files = TRUE,
  pp = FALSE,
  cores.left = NULL
)
}
\arguments{
\item{in.dir}{character. Directory containing the flagged HMD data}

\item{out.dir}{character. Directory where cleaned data should be saved}

\item{FF.remove}{NULL or character. Names of flags that should be removed. Default is NULL.
When NULL, the function will remove flags identified by the provider as bad flags (SPOOF_LOCATIONS, REPLAY).
When character, the function will remove the selected flags in addition to the default.}

\item{FF.suspect}{NULL or character. Names of flags that should be included as suspect. Default is NULL.
When NULL, the function will assign the following flags to suspect: OVER_CAPACITY_DEVICE, TOO_MANY_DEVICES_AT_LOCATION, HYPERV_OUTSIDE_CLUSTER, HYPERV_CLUSTER_INTERLEAVE, HYPERV_WITHIN_CLUSTER, and IMPLAUSIBLE_MOVEMENT or BI_LOCATION.}

\item{method}{character. What cleaning method should be used on the data? Many options exist. The default is 'none'. See details for other options.}

\item{all.keep}{logical. Should we keep all the data regardless of cleaning method? If a cleaning method is set to 'all' or an invalid cleaning method is selected, then this is overridden to TRUE}

\item{move.files}{logical. Should completed files be moved to a sub-directory in the the in.dir. This called "cleaned" by default.}

\item{pp}{logical. Should the function run using parallel processing?}

\item{cores.left}{numeric. How many cores should be reserved? Ignored when pp = FALSE.
Most functions should not be run on more than 8 cores. A warning is given when you use more than 8 cores.}
}
\value{
A list of file paths leading to the newly created cleaned data.
}
\description{
Remove locations containing bad data, locations with inaccurate fixes, or unreasonable movements
}
\details{
Accuracy is calculated based on company-provided values.
Accuracy is assigned a number from 0-5 where
0 = Unknown (no accuracy information from flags),
1 = Low accuracy (low accuracy flag (250 - 1000 m)),
2 = Medium-low accuracy (both low and medium accuracy flags (220 - 250 m)),
3 = Medium accuracy (medium accuracy flag (50 - 220 m)),
4 = Medium-high accuracy (both medium and high accuracy flags (35 - 50 m)),
and 5 = High accuracy (high accuracy flag (0 - 35 m)).

Methods for cleaning data: Many options are currently implemented for cleaning data depending on how conservative you need the data to be:
none: (Default) Only bad flags and duplicates will be removed.
all: Keeps all cleaning methods. Should only be used to compare the results of the various cleaning methods.
BaseHigh: Heather's original cleaning method which only removed high speeds (>190 km/h ~ > 120 mph) from suspect flags
BaseMed: A modification of the BaseHigh method removing lower speed locations (> 130 km/h ~ 80 mph) from suspect flags
HighSpeeds: Removes high speeds regardless of suspect flags
MedSpeeds: Removes medium speeds regardless of suspect flags
Accurate: Removes locations with unknown, low, or medium-low accuracy
SinglePoints: Removes locations where only one location in a day was detected
AccHigh: Removes locations that satisfies both the HighSpeeds and Accurate options
AccMed: Removes locations that satisfies both the MedSpeeds and Accurate options
HighSingle: Removes locations that satisfies both the HighSpeeds and SinglePoints options
MedSingle: Removes locations that satisfies both the MedSpeeds and SinglePoints options
AccHighSingle: Removes locations that satisfies the HighSpeeds, Accurate, and SinglePoints options
AccMedSingle: Removes locations that satisfies the MedSpeeds, Accurate, and SinglePoints options
}
\note{
There currently is no function in this package that creates the raw data...
}
\section{{Warning}}{

Working with HMD can be very memory intensive.
When using parallel processing, be careful how many cores you use. This can easily overuse RAM and crash the computer.
Functions that allow parallel processing will give a warning message if you use more than 8 cores.
}

\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\seealso{
\code{\link{flagAssignment}}
}
\concept{HMD}
\keyword{files}
\keyword{manip}
