% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateNE}
\alias{generateNE}
\title{generateNE}
\usage{
generateNE(HMDM.data, rules.supp = 0.1, rules.conf = 0.8,
  rules.minlen = 2, rules.maxlen = 2)
}
\arguments{
\item{HMDM.data}{data by generateHMDM function}

\item{rules.supp}{a numeric value for the minimal support of an item set (default : 0.1)}

\item{rules.conf}{a numeric value for the minimal confidence of rules/association hyperedges (default : 0.8)}

\item{rules.minlen}{an integer value for the minimal number of items per item set (default : 2)}

\item{rules.maxlen}{an integer value for the maximal number of items per item set (default : 2)}
}
\value{
Node and Edge object(data.frame) from apriori function
}
\description{
Create Node and Edge object from apriroi function
}
\examples{
# Set support value 0.01
generateNE(HMDM.data=data.freq, rules.supp=0.01)
}

