% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{loadGPKG}
\alias{loadGPKG}
\title{load tiled geopackages}
\usage{
loadGPKG(l, n)
}
\arguments{
\item{l}{list of geopackage file paths}

\item{n}{name of the geopackage to use. Defines names of columns for this data}
}
\value{
A spatial data.frame of vector data within tiles
}
\description{
Load in data from geopackages and subset to tiles
}
\details{
This function is intended to be used within the \code{\link{classifyHMD}} function. It is not currently set up to run standalone
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\references{
\code{\link{classifyHMD}}
}
\concept{hmd}
\concept{vector data}
\keyword{data}
