% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{loadRast}
\alias{loadRast}
\title{Load tiled raster data}
\usage{
loadRast(x, name)
}
\arguments{
\item{x}{Directory containing rasters}

\item{name}{Name to use for the raster layer}
}
\value{
A raster within the study area
}
\description{
Load in raster data within a study area
}
\details{
This function is intended to be used within the \code{\link{classifyHMD}} function. This function is standalone
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\concept{hmd}
\concept{vector data}
\keyword{data}
