% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{loadVector}
\alias{loadVector}
\title{Load in vector data}
\usage{
loadVector(x, gpkg)
}
\arguments{
\item{x}{character. Name of the layer. Need not be a full name but must be unique.}

\item{gpkg}{character. Name of the geopackage containing the data}
}
\value{
A spatial data frame within the study area
}
\description{
Load in vector data from a geopackage based on a short name of a layer
}
\details{
This function is intended to be used within the \code{\link{classifyHMD}} function. This function is standalone
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\concept{hmd}
\concept{vector data}
\keyword{data}
