% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readData}
\alias{readData}
\title{readData}
\usage{
readData(data.name, folder.path, pattern = "*.csv")
}
\arguments{
\item{data.name}{data name to be used. At least one vector}

\item{folder.path}{Folder path of data}

\item{pattern}{Pattern of load data}
}
\value{
data file
}
\description{
Use data folder path, read Data function
}
\examples{
# Two csv files exists 'C:/User/USER/Desktop/data' path
# One data is 'data1', the other is 'data2'

readData(data.name=c('data1','data2'), folder.path='C:/User/USER/Desktop/data', pattern='*.csv')
}

