% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_functions.R
\name{splitByMonth}
\alias{splitByMonth}
\title{Split HMD by month}
\usage{
splitByMonth(
  in.dir,
  out.dir,
  time.col = "timestamp_POSIXct.UTC",
  name.prefix,
  name.type
)
}
\arguments{
\item{in.dir}{character. The directory containing the file(s) to split up.}

\item{out.dir}{character. The directory where files should be saved.}

\item{time.col}{character. The column name of the time column used to determine months. Defaults to "timestamp_POSIXct.UTC" which is created by \code{\link{flagAssignment}}}

\item{name.prefix}{character. The name of this data that should be common among all newly created data}

\item{name.type}{character. A suffix for the data related to what stage of processing it is in. Usually one of (raw = "", flagged = "flagged", cleaned = "clean")}
}
\value{
A list of file paths to the newly created data
}
\description{
Split a combined HMD file by month
}
\details{
This function can help to make HMD more manageable and easier to process through other functions in this package, especially \code{\link{classifyHMD}}.
The file name of the new data will be in the format:
[name.prefix]_[first of month in YYYYMMDD]_[last of month in YYYYMMDD]_[name.type (usually one of c("", "flagged", "clean"))]_[date created in YYYYMMDD format].RDS
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\seealso{
\code{\link{flagAssignment}}, \code{\link{flagRemoval}}, \code{\link{classifyHMD}}
}
\concept{HMD}
\keyword{manip}
