% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_functions.R
\name{summarizeHMD}
\alias{summarizeHMD}
\title{Summarize HMD tracks}
\usage{
summarizeHMD(
  in.dir,
  out.dir,
  in.name,
  out.name,
  sort.cols,
  excl.type = c("bldg", "main", "rail"),
  add.calcs = TRUE,
  pp = FALSE,
  cores.left = NULL
)
}
\arguments{
\item{in.dir}{character. Directory containing the classified data to be summarized.}

\item{out.dir}{character. Directory where summarized data should be stored.}

\item{in.name}{character. String indicating how the classified data is defined. This should typically be "classified".}

\item{out.name}{character. String indicating what the summarized data should be called.}

\item{sort.cols}{character. Which columns should be used to define the summaries.}

\item{excl.type}{character. Which hmd.types should be excluded from the summaries. This defaults to c("bldg", "main", "rail").}

\item{add.calcs}{character. Were additional columns created in \code{\link{classifyHMD}} and should they be included in this function.}

\item{pp}{logical. Should the function run using parallel processing?}

\item{cores.left}{numeric. How many cores should be reserved? Ignored when pp = FALSE.
Most functions should not be run on more than 8 cores. A warning is given when you use more than 8 cores.}
}
\value{
A list of file paths leading to the newly created summarized data
}
\description{
Summarize tracks by different groups
}
\details{
This function is intended to summarize classified HMD as created by the \code{\link{classifyHMD}} function.
It summarizes data by user-selected columns as defined by the sort.cols argument. This should typically at minimum be grid (unique user ID) and day.
}
\section{{Unique metrics}}{

Metrics whose calculations are derived from the \code{\link[amt]{amt}} package:

(total distance) = (euclidean distance between first and last points)

(cumulative distance) = sum(step lengths)

tortuosity = (total distance) / (cumulative distance)
}

\section{{Warning}}{

Working with HMD can be very memory intensive.
When using parallel processing, be careful how many cores you use. This can easily overuse RAM and crash the computer.
Functions that allow parallel processing will give a warning message if you use more than 8 cores.
}

\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
# To estimate the sampling rate for each track:
dt[,("min" = min(difft_sec),
     "1st Qu." = quantile(difft_sec, probs = 0.25),
     "med" = median(difft_sec),
     "3rd Qu." = quantile(difft_sec, probs = 0.75),
     "max" = max(difft_sec),
     "mean" = mean(difft_sec),
     "sd" = sd(difft_sec),
     "NAs" = sum(is.na(difft_sec))),
    by = .(grid, day)]

## No real example right now
}
}
\references{
\code{\link{flagAssignment}}, \code{\link{flagRemoval}}, \code{\link{classifyHMD}} for functions in the HMD processing pipeline.
This function uses the \code{\link{numSummary}}, \code{\link{factSummary}} functions.
}
\concept{classification}
\concept{hmd}
\keyword{manip}
