% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_functions.R
\name{thinHMD}
\alias{thinHMD}
\title{Resample HMD to a set sampling rate}
\usage{
thinHMD(
  x,
  idcol,
  dtcol,
  rate = lubridate::seconds(30),
  tolerance = lubridate::seconds(10),
  thin = TRUE
)
}
\arguments{
\item{x}{data.table. Data.table containing the time information.}

\item{idcol}{character. Column name of the unique user id column}

\item{dtcol}{character. Column name of the date-time column}

\item{rate}{period or numeric. What is the expected sampling rate to include.
This can be either a period object from the lubridate package or a numeric in seconds.}

\item{tolerance}{period or numeric. The range tolerance for how close consecutive locations can be to exclude
This can be either a period object from the lubridate package or a numeric in seconds.}

\item{thin}{logical. Should rows be removed if they are too close together?}
}
\value{
One of:
When thin == FALSE: X with a new column added called burst. Bursts < 0 are locations that are sampled more frequently than is allowed by rate and tolerance
When thin == TRUE: X with extra locations removed.
}
\description{
Resample HMD locations to remove locations that are too close in time
}
\details{
This function is derived from and modified from the function amt::track_resample.
I adapted functions from amt to better accommodate running on multiple tracks.
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\references{
\code{\link[amt]{amt}}
}
\concept{hmd}
\concept{track resampling}
\keyword{manip}
