% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_functions.R
\name{trackFun}
\alias{trackFun}
\title{Movement metrics for HMD data}
\usage{
trackFun(ds, xcol = "X", ycol = "Y", dtcol = "ts_UTC", idcol = "grid")
}
\arguments{
\item{ds}{data.table. Data.table of cleaned or classified HMD to add movement metrics to}

\item{xcol}{character. Column name of the column storing x coordinates}

\item{ycol}{character. Column name of the column storing y coordinates}

\item{dtcol}{character. Column name of the column storing date-time information}

\item{idcol}{character. Column name of the column storing a unique user id}
}
\value{
The original ds with movement metrics added. Adds difft_sec, stepl_m, speed_mps, speed_kmh, direction, and ta.
Direction of travel assigns Eastward movements to 0.
}
\description{
Calculate movement metrics (speed, step length, and time between points) for each track
}
\details{
This function is modified from \code{\link[amt]{make_track}} which utilizes \code{\link[amt]{step_lengths}}, \code{\link[amt]{direction_rel}}, and \code{\link[amt]{speed}}.
If this function is run multiple times on a dataset, such as after running \code{\link{thinHMD}}, it will replace movement metrics each time.
All calculations are made forward (e.g., between fix 1 and fix 2)
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\references{
\code{\link[amt]{amt}} \code{\link[amt]{make_track}}
}
\seealso{
\code{\link{classifyHMD}}
}
\concept{HMD}
\keyword{manip}
