% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{calculateIndices}
\alias{calculateIndices}
\title{Create index across multiple files}
\usage{
calculateIndices(in.files, starting = 1)
}
\arguments{
\item{in.files.}{character vector of file paths to saved RDS files containing data.tables or data.frames.}

\item{starting.}{Numeric. What should the starting value of the index be? Defaults to 1}
}
\value{
data.table with index number of the in.files, the number of rows in each file, the starting index number, and ending index number
}
\description{
Calculate number of rows in each month to create a continuous OID across months
}
\details{
This function is standalone but intended to be used within the \code{\link{flagRemoval}} as a way to create a unique OID that starts at the beginning of the function.
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\references{
\code{\link{flagAssignment}}
}
\concept{hmd}
\concept{indicing}
\keyword{manip}
