% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_functions.R
\name{classifyHMD}
\alias{classifyHMD}
\title{Classify HMD based on distance to features}
\usage{
classifyHMD(
  in.dir,
  out.dir,
  studyarea,
  coord.sys,
  data.dir,
  ths,
  road.source,
  snowfraction = 50,
  add.calcs = TRUE,
  pp = FALSE,
  cores.left = NULL
)
}
\arguments{
\item{in.dir}{character. Directory containing cleaned HMD}

\item{out.dir}{character. Directory where classified HMD should be stored}

\item{studyarea}{Study area shapefile. This can be a character vector of a file path to a shapefile, "HMD" (Uses the area of the HMD to define a study area), or an sf or terra object.}

\item{coord.sys}{character. Projected coordinate system in epsg format.
If a geographic coordinate system is used (e.g., WGS84 or NAD83), this will default to epsg:3857.}

\item{data.dir}{character. Directory where the geopackage and raster layers are stored.
See details for important information about the structure of this directory}

\item{ths}{list. Distance thresholds for main roads (main), railroads (rail), buildings (bldg), local roads (local), and trails (trail).
If not all names are provided, this function will exclude those from determining hmd.type.}

\item{road.source}{character. The source data to use for local roads and trail data.
This should be any of ("NTD", "USFS").
If both are provided, then the function will use the nearest of the two to determine distance to local roads/trails.}

\item{snowfraction}{numeric between 0 and 100. What percentage of snow cover in a cell is required to assign a point as having snow.}

\item{add.calcs}{logical. Should the additional, non-distance calculations be included? See details for which calculations are included}

\item{pp}{logical. Should the function run using parallel processing?}

\item{cores.left}{numeric. How many cores should be reserved? Ignored when pp = FALSE}
}
\value{
A list of file paths leading to the newly created classified data.
}
\description{
Calculate spatial and track-level metrics from HMD data
}
\details{
This function uses a previously downloaded data directory containing spatial data and calculates distance to features and possibly additional spatial data associated with HMD.
Additional data currently being calculated is: the land manager, protected area status, distance to cell towers, elevation and terrain metrics (TRI, slope, aspect), snow cover, whether the point is in water, and whether the point is in an urban area.
}
\note{
While parallel processing is possible, this is not recommended for this function.
Even for moderately sized HMD files, distance calculations are very memory intensive and can easily over-use the system memory.
}
\section{{Warning}}{

Working with HMD can be very memory intensive.
When using parallel processing, be careful how many cores you use.
The function will give a warning message if you use more than 5 cores.
}

\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\seealso{
\code{\link{flagAssignment}}, \code{\link{flagRemoval}} for functions in HMD processing pipeline.
This function uses the \code{\link{trackFun}} and \code{\link{distToFeature}} functions.
}
\concept{Classification}
\concept{HMD}
\keyword{manip}
