% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{factSummary}
\alias{factSummary}
\title{Summarize factor data}
\usage{
factSummary(f, ds, by.cols = sort.cols)
}
\arguments{
\item{f}{character. Column name of the factor column that should be summarized.}

\item{ds}{data.table. The data.table that should be summarized.}

\item{by.cols}{character. Which columns should be used for grouping data.}
}
\value{
A data.table containing the by.cols and columns named after the factor of choice
}
\description{
Summarize factor data by specified columns
}
\details{
This function is intended to be used within the \code{\link{summarizeHMD}} function. It likely works standalone but is untested.
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\concept{hmd}
\concept{summary}
\keyword{manip}
