% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_functions.R
\name{flagAssignment}
\alias{flagAssignment}
\title{Assign flags to HMD}
\usage{
flagAssignment(
  in.dir,
  out.dir,
  move.file = TRUE,
  pp = FALSE,
  cores.left = NULL
)
}
\arguments{
\item{in.dir}{character. Directory containing the raw HMD data}

\item{out.dir}{character. Directory where flagged data should be saved}

\item{move.file}{logical. Should completed files be moved to a sub-directory in the the in.dir. This called "flagged" by default.}

\item{pp}{logical. Should the function run using parallel processing?}

\item{cores.left}{numeric. How many cores should be reserved? Ignored when pp = FALSE}
}
\value{
A list of file paths leading to the newly created flagged data.
}
\description{
Convert integer objects in Raw HMD data to readable numeric and character vectors and assign and name flags.
}
\details{
This function takes raw HMD and assigns flags and converts an integer date into a readable POSIXct date.
The local time is based on the timezone field in the raw data.
}
\note{
There currently is no function in this package that creates the raw data...
}
\section{{Warning}}{

Working with HMD can be very memory intensive.
When using parallel processing, be careful how many cores you use.
The function will give a warning message if you use more than 5 cores.
}

\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\seealso{
\code{\link{flagRemoval}}, \code{\link{classifyHMD}}
}
\concept{HMD}
\keyword{files}
\keyword{manip}
