% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd2-package.R
\docType{package}
\name{hmd-package}
\alias{hmd}
\alias{hmd-package}
\title{hmd: Process Human Mobility Data for Use in Wildlife Research}
\description{
This package was made to hold functions created by me to process human mobility data. It likely does not translate well to HMD from other sources and all functions are likely to be buggy.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/tomyamashita/hmd2/issues}
}

}
\author{
\strong{Maintainer}: Thomas J. Yamashita \email{tjyamashta@gmail.com} (\href{https://orcid.org/0000-0002-0213-6310}{ORCID})

Authors:
\itemize{
  \item Thomas J. Yamashita \email{tjyamashta@gmail.com} (\href{https://orcid.org/0000-0002-0213-6310}{ORCID})
}

}
\keyword{internal}
