% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{numSummary}
\alias{numSummary}
\title{Summarize numeric data}
\usage{
numSummary(f, ds, cols, by.cols = sort.cols)
}
\arguments{
\item{f}{character. Which type of numeric summary should be created.
This should be one of c("N", "mn", "min", "q01", "q05", "q25", "med", "q75", "q95" "q99", "max", "sd", "first" "last", "cum_dist", "cnt").}

\item{ds}{data.table. The data.table that should be summarized.}

\item{cols}{character. Which columns should be summarized.}

\item{by.cols}{character. Which columns should be used for grouping data.}
}
\value{
A data.table containing the by.cols and summarized cols
}
\description{
Summarize numeric data by specified columns
}
\details{
This function is intended to be used within the \code{\link{summarizeHMD}} function. It likely works standalone but is untested.
}
\section{{Disclaimer}}{

The functions in this package are in active development and likely has bugs.
Additionally, the functions in this package likely will not work properly if this package was not used to create the data.
}

\examples{
\dontrun{
## No example right now
}
}
\concept{hmd}
\concept{summary}
\keyword{manip}
