% Generated by roxygen2 (4.0.2): do not edit by hand
\name{splitLevel}
\alias{splitLevel}
\title{splitLevel}
\usage{
splitLevel(HMDM.data)
}
\arguments{
\item{HMDM.data}{data by generateHMDM function}
}
\value{
splited level
}
\description{
HMDM(Hierarchical Multi-Dictionary term Matrix) level split function
}
\examples{
#
sony.level <- split(HMDM.data=data.freq)
}

